+++ map/valid/test.cub
@@ -4,7 +4,7 @@ WE tex/blue_stone.xpm
 EA tex/blue_stone.xpm
 
 F 97,97,97
-C 1 0,20,50
+C 10,20,50
 
 11111
 1S001
